# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Classmate API specification
 * OpenAPI spec version: 0.1.0
 */
export interface Enrollment {
  id: number;
  courseId: number;
  studentId: number;
  enrolledAt: string;
  enrolledBy: number;
  isActive: boolean;
  droppedAt?: string | null;
  droppedBy?: number | null;
}

export interface EnrollStudentBody {
  studentId: number;
}

export interface HealthStatus {
  status: string;
}

export interface Student {
  id: number;
  name: string;
  email: string;
  grade: string;
  avatarUrl?: string | null;
  enrolledCourseIds: number[];
  createdAt: string;
}

export interface CreateStudentBody {
  name: string;
  email: string;
  grade: string;
  avatarUrl?: string | null;
}

export interface UpdateStudentBody {
  name?: string;
  email?: string;
  grade?: string;
  avatarUrl?: string;
}

export interface StudentProgress {
  studentId: number;
  totalAssignments: number;
  completedAssignments: number;
  averageScore: number;
  completionRate: number;
  topicsMastered: string[];
  topicsNeedingWork: string[];
}

export type CourseStatus = (typeof CourseStatus)[keyof typeof CourseStatus];

export const CourseStatus = {
  active: "active",
  archived: "archived",
} as const;

export interface Course {
  id: number;
  name: string;
  description: string;
  teacherName: string;
  teacherId?: number | null;
  subject: string;
  grade?: string | null;
  academicYear?: string | null;
  studentCount: number;
  status: CourseStatus;
  createdAt: string;
  updatedAt: string;
  createdBy?: number | null;
  updatedBy?: number | null;
}

export type CreateCourseBodyStatus =
  (typeof CreateCourseBodyStatus)[keyof typeof CreateCourseBodyStatus];

export const CreateCourseBodyStatus = {
  active: "active",
  archived: "archived",
} as const;

export interface CreateCourseBody {
  name: string;
  subject: string;
  grade: string;
  academicYear: string;
  teacherId: number;
  status?: CreateCourseBodyStatus;
  description?: string;
}

export type UpdateCourseBodyStatus =
  (typeof UpdateCourseBodyStatus)[keyof typeof UpdateCourseBodyStatus];

export const UpdateCourseBodyStatus = {
  active: "active",
  archived: "archived",
} as const;

export interface UpdateCourseBody {
  name?: string;
  subject?: string;
  grade?: string;
  academicYear?: string;
  teacherId?: number;
  status?: UpdateCourseBodyStatus;
  description?: string;
}

export type AssignmentStatus =
  (typeof AssignmentStatus)[keyof typeof AssignmentStatus];

export const AssignmentStatus = {
  pending: "pending",
  submitted: "submitted",
  graded: "graded",
  late: "late",
} as const;

export interface Assignment {
  id: number;
  title: string;
  description: string;
  courseId: number;
  courseName: string;
  studentId: number;
  studentName: string;
  dueDate: string;
  status: AssignmentStatus;
  score?: number | null;
  maxScore: number;
  feedback?: string | null;
  createdAt: string;
  updatedAt?: string;
  createdBy?: number | null;
  updatedBy?: number | null;
}

export interface CreateAssignmentBody {
  title: string;
  description: string;
  courseId: number;
  studentId: number;
  dueDate: string;
  maxScore: number;
}

export type UpdateAssignmentBodyStatus =
  (typeof UpdateAssignmentBodyStatus)[keyof typeof UpdateAssignmentBodyStatus];

export const UpdateAssignmentBodyStatus = {
  pending: "pending",
  submitted: "submitted",
  graded: "graded",
  late: "late",
} as const;

export interface UpdateAssignmentBody {
  status?: UpdateAssignmentBodyStatus;
  score?: number;
  feedback?: string;
}

export type AnnouncementPriority =
  (typeof AnnouncementPriority)[keyof typeof AnnouncementPriority];

export const AnnouncementPriority = {
  normal: "normal",
  urgent: "urgent",
} as const;

export interface Announcement {
  id: number;
  title: string;
  content: string;
  courseId: number;
  courseName: string;
  authorName: string;
  priority: AnnouncementPriority;
  createdAt: string;
  updatedAt?: string;
  createdBy?: number | null;
  updatedBy?: number | null;
}

export type AnnouncementInputPriority =
  (typeof AnnouncementInputPriority)[keyof typeof AnnouncementInputPriority];

export const AnnouncementInputPriority = {
  normal: "normal",
  urgent: "urgent",
} as const;

export interface AnnouncementInput {
  title: string;
  content: string;
  courseId: number;
  authorName: string;
  priority?: AnnouncementInputPriority;
}

export type AnnouncementUpdatePriority =
  (typeof AnnouncementUpdatePriority)[keyof typeof AnnouncementUpdatePriority];

export const AnnouncementUpdatePriority = {
  normal: "normal",
  urgent: "urgent",
} as const;

export interface AnnouncementUpdate {
  title?: string;
  content?: string;
  authorName?: string;
  priority?: AnnouncementUpdatePriority;
}

export interface Note {
  id: number;
  title: string;
  content: string;
  courseId: number;
  courseName: string;
  topic: string;
  videoUrl?: string | null;
  createdAt: string;
  updatedAt?: string;
  createdBy?: number | null;
  updatedBy?: number | null;
}

export interface CreateNoteBody {
  title: string;
  content: string;
  courseId: number;
  topic: string;
  videoUrl?: string | null;
}

export interface UpdateNoteBody {
  title?: string;
  content?: string;
  topic?: string;
  videoUrl?: string;
}

export interface Assessment {
  id: number;
  studentId: number;
  studentName: string;
  courseId: number;
  courseName: string;
  title: string;
  score: number;
  maxScore: number;
  percentage: number;
  strengths: string[];
  weaknesses: string[];
  createdAt: string;
  updatedAt?: string;
  createdBy?: number | null;
  updatedBy?: number | null;
}

export interface CreateAssessmentBody {
  studentId: number;
  courseId: number;
  title: string;
  score: number;
  maxScore: number;
  strengths: string[];
  weaknesses: string[];
}

export type SuggestionPriority =
  (typeof SuggestionPriority)[keyof typeof SuggestionPriority];

export const SuggestionPriority = {
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export interface Suggestion {
  area: string;
  priority: SuggestionPriority;
  suggestion: string;
  relatedTopic?: string | null;
}

export interface AiSuggestions {
  studentId: number;
  studentName: string;
  overallScore: number;
  suggestions: Suggestion[];
  recommendedTopics: string[];
  strengths: string[];
  generatedAt: string;
}

export interface StudentSummary {
  id: number;
  name: string;
  averageScore: number;
}

export interface DashboardSummary {
  totalStudents: number;
  totalCourses: number;
  totalAssignments: number;
  pendingAssignments: number;
  averageClassScore: number;
  completionRate: number;
  atRiskStudents: number;
  topPerformers: StudentSummary[];
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  assignment_submitted: "assignment_submitted",
  assignment_graded: "assignment_graded",
  note_created: "note_created",
  assessment_completed: "assessment_completed",
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  description: string;
  studentName: string;
  courseName: string;
  timestamp: string;
}

export interface GradeBreakdown {
  courseName: string;
  courseId: number;
  averageScore: number;
  aCount: number;
  bCount: number;
  cCount: number;
  dCount: number;
  fCount: number;
}

export interface LoginBody {
  username: string;
  password: string;
}

export interface AuthUser {
  id: number;
  username: string;
  displayName: string;
  role: string;
}

export type Logout200 = {
  ok: boolean;
};

export type ListAssignmentsParams = {
  courseId?: number;
  studentId?: number;
};

export type ListNotesParams = {
  courseId?: number;
};

export type ListAnnouncementsParams = {
  courseId?: number;
};

export type ListAssessmentsParams = {
  studentId?: number;
  courseId?: number;
};