# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Classmate API specification
servers:
  - url: /api
    description: Base API path
security:
  - sessionCookie: []
...
paths:
...
  /announcements:
    get:
      operationId: listAnnouncements
      tags: [announcements]
      summary: List course announcements
      parameters:
        - name: courseId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Announcements list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Announcement"
    post:
      operationId: createAnnouncement
      tags: [announcements]
      summary: Create a course announcement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
  /announcements/{id}:
    get:
      operationId: getAnnouncement
      tags: [announcements]
      summary: Get an announcement by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Announcement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
    patch:
      operationId: updateAnnouncement
      tags: [announcements]
      summary: Update an announcement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementUpdate"
      responses:
        "200":
          description: Updated announcement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"

  # ---- ASSESSMENTS ----
  /assessments:
    get:
      operationId: listAssessments
      tags: [assessments]
      summary: List assessments
      parameters:
        - name: studentId
          in: query
          required: false
          schema:
            type: integer
        - name: courseId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Assessments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Assessment"
    post:
      operationId: createAssessment
      tags: [assessments]
      summary: Submit an assessment result
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssessmentBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assessment"
  /assessments/{id}:
    get:
      operationId: getAssessment
      tags: [assessments]
      summary: Get assessment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Assessment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assessment"
  /assessments/{id}/ai-suggestions:
    get:
      operationId: getAiSuggestions
      tags: [assessments]
      summary: Get AI-powered improvement suggestions for a student based on assessment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: AI suggestions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiSuggestions"
  /students/{id}/ai-suggestions:
    get:
      operationId: getStudentAiSuggestions
      tags: [assessments]
      summary: Get AI improvement suggestions for a student across all courses
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: AI suggestions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiSuggestions"

  # ---- DASHBOARD ----
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: connect.sid
      description: Session cookie set by the server after login
  schemas:
...
    Announcement:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        courseId:
          type: integer
        courseName:
          type: string
        authorName:
          type: string
        priority:
          type: string
          enum: [normal, urgent]
        createdAt:
          type: string
        updatedAt:
          type: string
        createdBy:
          type: integer
          nullable: true
        updatedBy:
          type: integer
          nullable: true
      required: [id, title, content, courseId, courseName, authorName, priority, createdAt]
    AnnouncementInput:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        courseId:
          type: integer
        authorName:
          type: string
        priority:
          type: string
          enum: [normal, urgent]
      required: [title, content, courseId, authorName]
    AnnouncementUpdate:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        authorName:
          type: string
        priority:
          type: string
          enum: [normal, urgent]
...
    Assessment:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
        courseId:
          type: integer
        courseName:
          type: string
        title:
          type: string
        score:
          type: number
        maxScore:
          type: number
        percentage:
          type: number
        strengths:
          type: array
          items:
            type: string
        weaknesses:
          type: array
          items:
            type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        createdBy:
          type: integer
          nullable: true
        updatedBy:
          type: integer
          nullable: true
      required: [id, studentId, studentName, courseId, courseName, title, score, maxScore, percentage, strengths, weaknesses, createdAt]
    CreateAssessmentBody:
      type: object
      properties:
        studentId:
          type: integer
        courseId:
          type: integer
        title:
          type: string
        score:
          type: number
        maxScore:
          type: number
        strengths:
          type: array
          items:
            type: string
        weaknesses:
          type: array
          items:
            type: string
      required: [studentId, courseId, title, score, maxScore, strengths, weaknesses]
    AiSuggestions:
      type: object
      properties:
        studentId:
          type: integer
        studentName:
          type: string
        overallScore:
          type: number
        suggestions:
          type: array
          items:
            $ref: "#/components/schemas/Suggestion"
        recommendedTopics:
          type: array
          items:
            type: string
        strengths:
          type: array
          items:
            type: string
        generatedAt:
          type: string
      required: [studentId, studentName, overallScore, suggestions, recommendedTopics, strengths, generatedAt]
    Suggestion:
      type: object
      properties:
        area:
          type: string
        priority:
          type: string
          enum: [high, medium, low]
        suggestion:
          type: string
        relatedTopic:
          type: string
          nullable: true
      required: [area, priority, suggestion]
...